// sdk/src/types.ts
export type Utxo = {
  txid: Buffer;
  index: number;
  value: number;
};
export type Network = 'mainnet' | 'testnet' | 'regtest';

/**
 * @see https://www.w3.org/TR/did-core/#verification-methods
 */
export type VerificationMethod = {
  /**
   * A multikey that specificies a public key for the verification method.
   * @see https://www.w3.org/TR/vc-data-integrity/#multikey
   */
  multikey: Buffer;
  /**  The verification relationships for this verification method. */
  verificationRelationshipFlags: VerificationRelationshipFlags;
};

export type Did = {
  /**
   * The verification methods for the DID.
   * @see https://www.w3.org/TR/did-core/#verification-methods
   */
  verificationMethods: VerificationMethod[];
  /** The p2tr key corresponding to the DID UTXO that can be used to update or deactivate the DID. */
  controllerKey?: Buffer;
  /** Additional metadata pertaining to the DID such as services. */
  metadata?: Record<string, object>;
  isDeactivated?: boolean;
};
// ...
export type VerificationMethodOperation =
  | VerificationMethodDeletion
  | VerificationMethodUpdate
  | VerificationMethodAppend;

/**
 * @see https://microstrategy.github.io/did-btc-spec/#update
 */
export type DidUpdate = {
  /** Changes to the verification methods for the DID. */
  vm?: VerificationMethodOperation[];
  /** Metadata records to update in the DID document */
  u?: Record<string, object>;
  /** Metadata to delete from the DID document. */
  d?: Record<string, object>;
  /** Metadata to add to the DID document. */
  a?: Record<string, object>;
};

export type WalletUtxo = {
  /**
   * A reference to the unspent transaction output (UTXO) to be spent for this transaction. This
   * must be a pay-to-taproot (P2TR) output whose internal public key corresponds to the private key
   * provided.
   * @see https://github.com/bitcoin/bips/blob/master/bip-0341.mediawiki#user-content-Taproot_key_path_spending_signature_validation
   */
  utxo: Utxo;
  /**
   * A 32 byte private key that corresponds to the internal public key of the UTXO being spent and
   * can therefore be used to produce a valid signature for the transaction.
   */
  privkey: Uint8Array;
};
// ...
export type BatchDidUpdate = {
  /** The current state of the DID that is being updated. */
  did: Did;
  /** The did index in its original batched creation transaction. */
  i: number;
  /**
   * The update operation to apply to the DID. If this is an empty object, the DID will be
   * deactivated.
   */
  update: DidUpdate;
};